segmented JSON data handling
